Added support for excluding files from revel build
This PR is in response to https://github.com/revel/revel/issues/1047.

This rudimentary implementation allows a developer to place glob file
patterns in a `.revelignore` file in the root directory of their revel
project. All project files which match any pattern in the file will be
exluced from all `revel build` and `revel package` output.